requirements.txt
beautifulsoup4==4.11.1
requests==2.28.1
pycryptodome==3.15.0